Fix bug when tracing only preemptions
The "else" block in pick_cpu() was leading
to infinite loop when num_cpus == 1 (only
preemptions). I've added a check to
prevent that.
Some minor things were also fixed, such as
usage() and the count of samples of preemptions
and migrations. When num_cpus == 1 we don't need
to check migration_counter.